requirements.txt
numpy ~= 1.26.1
pandas ~= 2.2.1
matplotlib ~= 3.8.1
plotly ~= 5.21.0
kaleido ~= 0.2.1
wordcloud ~= 1.9.3
requests ~= 2.31.0
spacy ~= 3.7.1
scikit-learn == 1.4.2
tokenizers == 0.19.1
sentencepiece == 0.2.0
datasets == 2.19.0
torch == 2.2.1
transformers == 4.40.0
evaluate == 0.4.1
tiktoken == 0.6.0
dgl == 2.1.0
openai == 1.22.0
mlflow == 2.12.1
tqdm ~= 4.66.1
click ~= 8.1.7
python-dotenv ~= 1.0.1
